restructured staff management feature
