+++ backend/kpi-tracker-services/src/main/java/org/pahappa/systems/kpiTracker/core/services/systemUsers/StaffService.java
@@ -2,7 +2,7 @@ package org.pahappa.systems.kpiTracker.core.services.systemUsers;
 
 import org.pahappa.systems.kpiTracker.core.services.GenericService;
 import org.pahappa.systems.kpiTracker.models.organization_structure.Department;
-import org.pahappa.systems.kpiTracker.models.systemUsers.Staff;
+import org.pahappa.systems.kpiTracker.models.staff.Staff;
 
 import java.util.List;
